assignment1/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
idna==3.4
importlib-resources==6.4.5
jinja2==3.1.4
kiwisolver==1.4.7
lightning-utilities==0.11.9
markupsafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.0
numpy==1.24.1
packaging==24.2
pillow==10.2.0
propcache==0.2.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pyyaml==6.0.2
requests==2.28.1
six==1.17.0
sympy==1.13.1
torchmetrics==1.5.2
tqdm==4.67.1
triton==2.1.0
typing-extensions==4.12.2
urllib3==1.26.13
yarl==1.15.2
zipp==3.20.2